requirements.txt
fastapi
snowflake-snowpark-python
uvicorn
